requirements.txt
# Core runtime deps
requests>=2.31.0
PyYAML>=6.0.1
openai
